@@ -22,6 +22,7 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.Reader;
+import java.net.URI;
 import java.net.URL;
 import java.util.Collections;
 import java.util.List;
@@ -56,7 +57,7 @@ public class PythonModuleParser extends PythonParser<ModuleEntry> {
   }
 
   public static void main(String[] args) throws Exception {
-    URL url = new URL(args[0]);
+    URL url = new URI(args[0]).toURL();
     List<File> pythonPath = Util.getPathFiles(args[1]);
 
     PythonParser<ModuleEntry> p =
@@ -67,7 +68,7 @@ public class PythonModuleParser extends PythonParser<ModuleEntry> {
             pythonPath);
     CAstEntity script = p.translateToCAst();
     System.err.println(script);
-    System.err.println(CAstPrinter.print(script));
+    System.err.println(CAstPrinter.print(script.getAST()));
   }
 
   @Override
