@@ -7,7 +7,7 @@ import com.ibm.wala.cast.python.client.PythonAnalysisEngine;
 import com.ibm.wala.cast.python.ml.analysis.TensorTypeAnalysis;
 import com.ibm.wala.cast.python.ml.analysis.TensorVariable;
 import com.ibm.wala.cast.python.ml.client.PythonTensorAnalysisEngine;
-import com.ibm.wala.cast.python.test.TestPythonCallGraphShape;
+import com.ibm.wala.cast.python.test.TestJythonCallGraphShape;
 import com.ibm.wala.cast.python.types.PythonTypes;
 import com.ibm.wala.cast.python.util.Util;
 import com.ibm.wala.cast.types.AstMethodReference;
@@ -29,13 +29,14 @@ import com.ibm.wala.util.NullProgressMonitor;
 import com.ibm.wala.util.collections.HashSetFactory;
 import java.io.File;
 import java.io.IOException;
-import java.net.URL;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
 
-public abstract class TestPythonMLCallGraphShape extends TestPythonCallGraphShape {
+public abstract class TestPythonMLCallGraphShape extends TestJythonCallGraphShape {
 
   @FunctionalInterface
   protected interface CheckTensorOps {
@@ -109,9 +110,9 @@ public abstract class TestPythonMLCallGraphShape extends TestPythonCallGraphShap
   }
 
   protected void checkTensorOps(String url, CheckTensorOps check)
-      throws IllegalArgumentException, CancelException, IOException {
+      throws IllegalArgumentException, CancelException, IOException, URISyntaxException {
     PythonTensorAnalysisEngine e = new PythonTensorAnalysisEngine();
-    e.setModuleFiles(Collections.singleton(new SourceURLModule(new URL(url))));
+    e.setModuleFiles(Collections.singleton(new SourceURLModule(new URI(url).toURL())));
     PropagationCallGraphBuilder cgBuilder =
         (PropagationCallGraphBuilder) e.defaultCallGraphBuilder();
     CallGraph CG = cgBuilder.makeCallGraph(cgBuilder.getOptions());
