@@ -2,12 +2,8 @@ package com.ibm.wala.cast.python.test;
 
 import static java.util.Collections.emptyList;
 
-import com.ibm.wala.cast.ipa.callgraph.CAstCallGraphUtil;
 import com.ibm.wala.cast.python.client.PythonAnalysisEngine;
-import com.ibm.wala.cast.python.loader.PythonLoaderFactory;
 import com.ibm.wala.cast.python.types.PythonTypes;
-import com.ibm.wala.cast.python.util.PythonInterpreter;
-import com.ibm.wala.cast.python.util.Util;
 import com.ibm.wala.cast.types.AstMethodReference;
 import com.ibm.wala.cast.util.test.TestCallGraphShape;
 import com.ibm.wala.classLoader.Module;
@@ -15,20 +11,18 @@ import com.ibm.wala.classLoader.SourceURLModule;
 import com.ibm.wala.core.util.strings.Atom;
 import com.ibm.wala.ipa.callgraph.CGNode;
 import com.ibm.wala.ipa.callgraph.CallGraph;
-import com.ibm.wala.ipa.callgraph.CallGraphBuilder;
-import com.ibm.wala.ipa.callgraph.propagation.InstanceKey;
 import com.ibm.wala.ipa.callgraph.propagation.PropagationCallGraphBuilder;
-import com.ibm.wala.ipa.callgraph.propagation.SSAPropagationCallGraphBuilder;
 import com.ibm.wala.ipa.cha.ClassHierarchyException;
 import com.ibm.wala.types.MethodReference;
 import com.ibm.wala.types.TypeName;
 import com.ibm.wala.types.TypeReference;
 import com.ibm.wala.util.CancelException;
-import com.ibm.wala.util.NullProgressMonitor;
 import com.ibm.wala.util.collections.HashSetFactory;
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.URL;
 import java.util.Collection;
 import java.util.List;
@@ -36,22 +30,8 @@ import java.util.Set;
 
 public abstract class TestPythonCallGraphShape extends TestCallGraphShape {
 
-  static {
-    try {
-      Class<?> j3 = Class.forName("com.ibm.wala.cast.python.loader.Python3LoaderFactory");
-      PythonAnalysisEngine.setLoaderFactory((Class<? extends PythonLoaderFactory>) j3);
-      Class<?> i3 = Class.forName("com.ibm.wala.cast.python.util.Python3Interpreter");
-      PythonInterpreter.setInterpreter((PythonInterpreter) i3.newInstance());
-    } catch (ClassNotFoundException | InstantiationException | IllegalAccessException e) {
-      try {
-        Class<?> j2 = Class.forName("com.ibm.wala.cast.python.loader.Python2LoaderFactory");
-        PythonAnalysisEngine.setLoaderFactory((Class<? extends PythonLoaderFactory>) j2);
-        Class<?> i2 = Class.forName("com.ibm.wala.cast.python.util.Python2Interpreter");
-        PythonInterpreter.setInterpreter((PythonInterpreter) i2.newInstance());
-      } catch (ClassNotFoundException | InstantiationException | IllegalAccessException e1) {
-        assert false : e.getMessage() + ", then " + e1.getMessage();
-      }
-    }
+  public TestPythonCallGraphShape() {
+    super();
   }
 
   @Override
@@ -80,10 +60,10 @@ public abstract class TestPythonCallGraphShape extends TestCallGraphShape {
       if (f.exists()) {
         return new SourceURLModule(f.toURI().toURL());
       } else {
-        URL url = new URL(name);
+        URL url = new URI(name).toURL();
         return new SourceURLModule(url);
       }
-    } catch (MalformedURLException e) {
+    } catch (MalformedURLException | URISyntaxException | IllegalArgumentException e) {
       return new SourceURLModule(getClass().getClassLoader().getResource(name));
     }
   }
@@ -135,21 +115,5 @@ public abstract class TestPythonCallGraphShape extends TestCallGraphShape {
       sb.append(n.getIR()).append("\n");
     }
     return sb;
-  }
-
-  public static void main(String[] args)
-      throws ClassHierarchyException, IllegalArgumentException, CancelException, IOException {
-    TestPythonCallGraphShape driver = new TestPythonCallGraphShape() {};
-
-    PythonAnalysisEngine<?> E = driver.makeEngine(Util.getPathFiles(args[1]), args[0]);
-
-    CallGraphBuilder<? super InstanceKey> builder = E.defaultCallGraphBuilder();
-    CallGraph CG = builder.makeCallGraph(E.getOptions(), new NullProgressMonitor());
-
-    CAstCallGraphUtil.AVOID_DUMP.set(false);
-    CAstCallGraphUtil.dumpCG(
-        ((SSAPropagationCallGraphBuilder) builder).getCFAContextInterpreter(),
-        builder.getPointerAnalysis(),
-        CG);
   }
 }
