@@ -4,16 +4,12 @@ import org.python.util.PythonInterpreter;
 
 public class Python3Interpreter extends com.ibm.wala.cast.python.util.PythonInterpreter {
 
-  private static PythonInterpreter interp = null;
+  private static PythonInterpreter interp;
 
   public static PythonInterpreter getInterp() {
-    try {
-      if (interp == null) {
-        //			PySystemState.initialize(  );
-        interp = new PythonInterpreter();
-      }
-    } catch (Throwable e) {
-
+    if (interp == null) {
+      //			PySystemState.initialize(  );
+      interp = new PythonInterpreter();
     }
     return interp;
   }
