@@ -20,6 +20,7 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.PipedInputStream;
 import java.io.PipedOutputStream;
+import java.lang.reflect.InvocationTargetException;
 import java.net.Socket;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -42,17 +43,37 @@ public class ServerTest {
 
   static {
     try {
-      Class<?> j3 = Class.forName("com.ibm.wala.cast.python.loader.Python3LoaderFactory");
-      PythonAnalysisEngine.setLoaderFactory((Class<? extends PythonLoaderFactory>) j3);
+      @SuppressWarnings("unchecked")
+      Class<? extends PythonLoaderFactory> j3 =
+          (Class<? extends PythonLoaderFactory>)
+              Class.forName("com.ibm.wala.cast.python.loader.Python3LoaderFactory");
+      PythonAnalysisEngine.setLoaderFactory(j3);
       Class<?> i3 = Class.forName("com.ibm.wala.cast.python.util.Python3Interpreter");
-      PythonInterpreter.setInterpreter((PythonInterpreter) i3.newInstance());
-    } catch (ClassNotFoundException | InstantiationException | IllegalAccessException e) {
+      PythonInterpreter.setInterpreter(
+          (PythonInterpreter) i3.getDeclaredConstructor().newInstance());
+    } catch (ClassNotFoundException
+        | InstantiationException
+        | IllegalAccessException
+        | IllegalArgumentException
+        | InvocationTargetException
+        | NoSuchMethodException
+        | SecurityException e) {
       try {
-        Class<?> j2 = Class.forName("com.ibm.wala.cast.python.loader.Python2LoaderFactory");
-        PythonAnalysisEngine.setLoaderFactory((Class<? extends PythonLoaderFactory>) j2);
+        @SuppressWarnings("unchecked")
+        Class<? extends PythonLoaderFactory> j2 =
+            (Class<? extends PythonLoaderFactory>)
+                Class.forName("com.ibm.wala.cast.python.loader.Python2LoaderFactory");
+        PythonAnalysisEngine.setLoaderFactory(j2);
         Class<?> i2 = Class.forName("com.ibm.wala.cast.python.util.Python2Interpreter");
-        PythonInterpreter.setInterpreter((PythonInterpreter) i2.newInstance());
-      } catch (ClassNotFoundException | InstantiationException | IllegalAccessException e1) {
+        PythonInterpreter.setInterpreter(
+            (PythonInterpreter) i2.getDeclaredConstructor().newInstance());
+      } catch (ClassNotFoundException
+          | InstantiationException
+          | IllegalAccessException
+          | IllegalArgumentException
+          | InvocationTargetException
+          | NoSuchMethodException
+          | SecurityException e1) {
         assert false : e.getMessage() + ", then " + e1.getMessage();
       }
     }
