@@ -1,4 +1,4 @@
-""" Convolutional Neural Network.
+"""Convolutional Neural Network.
 
 Build and train a convolutional neural network with TensorFlow.
 This example is using the MNIST database of handwritten digits
