@@ -1,16 +1,14 @@
 package com.ibm.wala.cast.python.test;
 
-import com.ibm.wala.cast.ipa.callgraph.CAstCallGraphUtil;
 import com.ibm.wala.cast.python.client.PythonAnalysisEngine;
 import com.ibm.wala.cast.python.ipa.callgraph.PythonSSAPropagationCallGraphBuilder;
 import com.ibm.wala.ipa.callgraph.CallGraph;
-import com.ibm.wala.ipa.callgraph.propagation.SSAContextInterpreter;
 import com.ibm.wala.ipa.cha.ClassHierarchyException;
 import com.ibm.wala.util.CancelException;
 import java.io.IOException;
 import org.junit.Test;
 
-public class TestAssign extends TestPythonCallGraphShape {
+public class TestAssign extends TestJythonCallGraphShape {
 
   protected static final Object[][] assertionsAssign1 =
       new Object[][] {
@@ -47,9 +45,5 @@ public class TestAssign extends TestPythonCallGraphShape {
     PythonSSAPropagationCallGraphBuilder B = E.defaultCallGraphBuilder();
     CallGraph CG = B.makeCallGraph(B.getOptions());
     verifyGraphAssertions(CG, assertionsAssign2);
-
-    CAstCallGraphUtil.AVOID_DUMP.set(false);
-    CAstCallGraphUtil.dumpCG(
-        (SSAContextInterpreter) B.getContextInterpreter(), B.getPointerAnalysis(), CG);
   }
 }
