@@ -6,7 +6,7 @@ import com.ibm.wala.util.CancelException;
 import java.io.IOException;
 import org.junit.Test;
 
-public class TestCollections extends TestPythonCallGraphShape {
+public class TestCollections extends TestJythonCallGraphShape {
 
   @Test
   public void testCollections1()
