@@ -18,7 +18,6 @@ import com.ibm.wala.cast.python.ipa.summaries.PythonSuper;
 import com.ibm.wala.cast.python.ir.PythonLanguage;
 import com.ibm.wala.cast.python.loader.PythonLoaderFactory;
 import com.ibm.wala.cast.python.types.PythonTypes;
-import com.ibm.wala.cast.python.util.PythonInterpreter;
 import com.ibm.wala.cast.types.AstMethodReference;
 import com.ibm.wala.cast.util.Util;
 import com.ibm.wala.classLoader.IClass;
@@ -92,18 +91,28 @@ public abstract class PythonAnalysisEngine<T>
 
   static {
     try {
-      Class<?> j3 = Class.forName("com.ibm.wala.cast.python.loader.Python3LoaderFactory");
-      PythonAnalysisEngine.setLoaderFactory((Class<? extends PythonLoaderFactory>) j3);
-      Class<?> i3 = Class.forName("com.ibm.wala.cast.python.util.Python3Interpreter");
-      PythonInterpreter.setInterpreter((PythonInterpreter) i3.newInstance());
-    } catch (ClassNotFoundException | InstantiationException | IllegalAccessException e) {
+      @SuppressWarnings("unchecked")
+      Class<PythonLoaderFactory> j4 =
+          (Class<PythonLoaderFactory>)
+              Class.forName("com.ibm.wala.cast.python.loader.JepPythonLoaderFactory");
+      PythonAnalysisEngine.setLoaderFactory(j4);
+    } catch (UnsatisfiedLinkError | ClassNotFoundException e2) {
       try {
-        Class<?> j2 = Class.forName("com.ibm.wala.cast.python.loader.Python2LoaderFactory");
-        PythonAnalysisEngine.setLoaderFactory((Class<? extends PythonLoaderFactory>) j2);
-        Class<?> i2 = Class.forName("com.ibm.wala.cast.python.util.Python2Interpreter");
-        PythonInterpreter.setInterpreter((PythonInterpreter) i2.newInstance());
-      } catch (ClassNotFoundException | InstantiationException | IllegalAccessException e1) {
-        assert false : e.getMessage() + ", then " + e1.getMessage();
+        @SuppressWarnings("unchecked")
+        Class<? extends PythonLoaderFactory> j3 =
+            (Class<? extends PythonLoaderFactory>)
+                Class.forName("com.ibm.wala.cast.python.loader.Python3LoaderFactory");
+        PythonAnalysisEngine.setLoaderFactory(j3);
+      } catch (ClassNotFoundException e) {
+        try {
+          @SuppressWarnings("unchecked")
+          Class<? extends PythonLoaderFactory> j2 =
+              (Class<? extends PythonLoaderFactory>)
+                  Class.forName("com.ibm.wala.cast.python.loader.Python2LoaderFactory");
+          PythonAnalysisEngine.setLoaderFactory(j2);
+        } catch (ClassNotFoundException e1) {
+          assert false : e.getMessage() + ", then " + e1.getMessage();
+        }
       }
     }
   }
@@ -111,6 +120,7 @@ public abstract class PythonAnalysisEngine<T>
   private static Class<? extends PythonLoaderFactory> loaders;
 
   public static void setLoaderFactory(Class<? extends PythonLoaderFactory> lf) {
+    assert loaders == null;
     loaders = lf;
   }
 
